feat(alb-target): fix issue with listing alb targets
